backend/requirements.txt
djangorestframework-simplejwt==5.3.1
django-comment-migrate==0.1.5
# mysqlclient==2.1.0
mysqlclient
django-cors-headers==3.11.0
django-filter==21.1
django-typomatic==1.6.1
openpyxl_image_loader
django-import-export